Refactor: PaymentMethod and ShippingInfo have their own directories inside of the checkout page
